# core/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A high-performance proxy rotation server with health monitoring and intelligent routing\nProvides comprehensive API for managing proxy servers, monitoring their health,\nand configuring rotation strategies.",
        "title": "Rota Proxy API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/alpkeskin/rota"
        },
        "license": {
            "name": "LICENSE",
            "url": "https://github.com/alpkeskin/rota/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8001",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticate user and receive JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/dashboard/charts/response-time": {
            "get": {
                "description": "Get response time chart data for visualization",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Response time chart",
                "parameters": [
                    {
                        "type": "string",
                        "default": "4h",
                        "description": "Time interval (e.g., 4h, 24h)",
                        "name": "interval",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chart data",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ResponseTimeChartData"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/dashboard/charts/success-rate": {
            "get": {
                "description": "Get success rate chart data for visualization",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Success rate chart",
                "parameters": [
                    {
                        "type": "string",
                        "default": "4h",
                        "description": "Time interval (e.g., 4h, 24h)",
                        "name": "interval",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chart data",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.SuccessRateChartData"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/dashboard/stats": {
            "get": {
                "description": "Get dashboard statistics including proxy and request metrics",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Dashboard statistics",
                "responses": {
                    "200": {
                        "description": "Dashboard statistics",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.DashboardStats"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/logs": {
            "get": {
                "description": "Get paginated list of system logs with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "List logs",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by log level",
                        "name": "level",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by source",
                        "name": "source",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start time (RFC3339)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time (RFC3339)",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of logs",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.LogListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/logs/export": {
            "get": {
                "description": "Export system logs in various formats (txt, json)",
                "produces": [
                    "text/plain",
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Export logs",
                "parameters": [
                    {
                        "type": "string",
                        "default": "txt",
                        "description": "Export format (txt/json)",
                        "name": "format",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by log level",
                        "name": "level",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by source",
                        "name": "source",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start time (RFC3339)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time (RFC3339)",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exported file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/proxies": {
            "get": {
                "description": "Get paginated list of proxies with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proxies"
                ],
                "summary": "List proxies",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by protocol",
                        "name": "protocol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (asc/desc)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of proxies",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ProxyListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new proxy server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proxies"
                ],
                "summary": "Create proxy",
                "parameters": [
                    {
                        "description": "Proxy details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.CreateProxyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created proxy",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.Proxy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/proxies/bulk": {
            "post": {
                "description": "Create multiple proxy servers at once",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proxies"
                ],
                "summary": "Bulk create proxies",
                "parameters": [
                    {
                        "description": "List of proxies to create",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.BulkCreateProxyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Creation results",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/proxies/bulk-delete": {
            "post": {
                "description": "Delete multiple proxy servers at once",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proxies"
                ],
                "summary": "Bulk delete proxies",
                "parameters": [
                    {
                        "description": "List of proxy IDs to delete",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.BulkDeleteProxyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deletion results",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/proxies/export": {
            "get": {
                "description": "Export proxy list in various formats (txt, json, csv)",
                "produces": [
                    "text/plain",
                    "application/json",
                    "text/csv"
                ],
                "tags": [
                    "proxies"
                ],
                "summary": "Export proxies",
                "parameters": [
                    {
                        "type": "string",
                        "default": "txt",
                        "description": "Export format (txt/json/csv)",
                        "name": "format",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exported file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/settings/reset": {
            "post": {
                "description": "Reset configuration to default values",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Reset settings",
                "responses": {
                    "200": {
                        "description": "Reset confirmation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_alpkeskin_rota_core_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_alpkeskin_rota_core_internal_models.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_alpkeskin_rota_core_internal_models.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_alpkeskin_rota_core_internal_models.UserInfoResponse"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}